COMPLETE STAGE 1/2 Tests in progress

+++ app/Http/Controllers/CommentController.php
@@ -12,7 +12,6 @@ class CommentController extends Controller
 {
     public function index()
     {
-        $Comments = Comment::orderBy('created_at', 'desc')->paginate(20);
      }
     
 
@@ -29,23 +28,27 @@ class CommentController extends Controller
 
     /**
      * Store a newly created resource in storage.
-     *
+     *ds
      * @param  \Illuminate\Http\Request  $request
      * @return \Illuminate\Http\Response
      */
     public function store(Request $request,Post $post){
-            $request->validate([
-            'commment_text' => 'required']);
+        $request->validate([
+            'comment_text' => 'required|max:1000']);
 
-            $comment = new Comment;
-            $comment->user_id = Auth()->user()->id;
-            $comment->post_id = $request->input('post_id');
-            $comment->commment_text = $request->commment_text;
-            $comment->concern = rand(0,1);
-            $comment->save();
-            return redirect()->back()->with('success', 'New Comment Added.');
+        $comment = new Comment;
+        $comment->user_id = Auth()->user()->id;
+        $comment->post_id = $request->input('post_id');
+        $comment->comment_text = $request->comment_text;
+        $comment->concern = rand(0,1);
+        $comment->save();
+        $user = Auth()->user();
+        if($comment->concern == 1)
+            $user->concerns += 1;
+            $user->save();
+        return redirect()->back()->with('success', 'New Comment Added.');
 
-        }
+    }
 
     /**
      * Display the specified resource.
@@ -54,6 +57,7 @@ class CommentController extends Controller
      * @return \Illuminate\Http\Response
      */
     public function show(Comment $comment) {
+        
         return view('comments.show', [
         'comment' => $comment
         ]);
@@ -61,35 +65,69 @@ class CommentController extends Controller
 
         //edits comments
     public function edit(Comment $comment) {
-        return view('comments.edit', [
-        'comment' => $comment
-        ]);
+        if(auth()->user()->id == $comment->user_id){
+            return view('comments.edit', [
+            'comment' => $comment
+            ]);
+        } else{
+            abort(403, 'Unauthorized access');
         }
 
-    /**
-     * Update the specified resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @param  \App\Models\Comment  $comment
-     * @return \Illuminate\Http\Response
-     */
-    public function update(Request $request, Comment $comment) {
-        $request->validate([
-            'commment_text' => 'required']);
+        }
 
-        $comment->user_id = Auth()->user()->id;
-        $comment->commment_text = $request->commment_text;
-        $comment->concern = rand(0,1);
-        $comment->save();
-        return redirect()->route('posts.show', $comment->post_id)->with('success', 'Update Successfull');
-        
+    public function update(Request $request, Comment $comment) {
+        if(auth()->user()->id == $comment->user_id){
+            $request->validate([
+                'comment_text' => 'required|max:1000']);
+
+            $preconcern = $comment->concern;
+            $comment->user_id = Auth()->user()->id;
+            $comment->comment_text = $request->comment_text;
+            $comment->concern = rand(0,1);
+            $comment->save();
+
+            $user = Auth()->user();
+            if($preconcern < $comment->concern)
+                $user->concerns = $user->concerns += 1;
+            elseif($preconcern > $comment->concern && $user->concerns > 0){
+                $user->concerns = $user->concerns -= 1;
+            } else{
+                $user->concerns + 0;
+            }
+            $user->save();
+            return redirect()->route('posts.show', $comment->post_id)->with('success', 'Update Successfull');
+        } else{
+            abort(403, 'Unauthorized access');
+        }
     }
 
 
 
 
     public function destroy(Comment $comment) {
-        $comment->delete();
-        return redirect()->back()->with('success', 'New Comment Added.');
-        }
+        if(auth()->user()->id == $comment->user_id){
+            if(Auth()->user()->isAdmin == 1){
+                $user = User::find($comment->user_id);
+                if($comment->concern == 1 && $user->concerns > 0){
+                    $user->concerns = $user->concerns -= 1;
+                }
+                $user->reported -= $comment->comment_reports;
+                $comment->delete();
+                $user->save();
+                return redirect()->back()->with('success', 'Comment Deleted Successfully');
+            }
+            else{
+                $user = Auth()->user();
+                if($comment->concern == 1 && $user->concerns > 0){
+                    $user->concerns = $user->concerns -= 1;
+                }
+                $user->reported -= $comment->comment_reports;
+                $comment->delete();
+                $user->save();
+                return redirect()->route('posts.show', $comment->post_id)->with('success', 'Comment Deleted Successfully');
+            }
+        } else{
+            abort(403, 'Unauthorized access');
+        }   
+    }
 }
