+++ app/Http/Controllers/PostController.php
@@ -11,24 +11,29 @@ class PostController extends Controller
 {
 
     public function index(){
-        $posts = Post::orderBy('created_at', 'desc')->paginate(20);
+        $blocked_users = User::where('blocked', 1)->pluck('id');
+        $posts = Post::orderBy('created_at', 'desc')->whereNotIn('user_id', $blocked_users)->paginate(20);
         return view('posts.index', compact('posts'));
         
     }
     public function show(Post $post) {
-        $comments = Comment::where('post_id', $post->id)->orderBy('created_at', 'desc')->paginate(20);
+        $blocked_users = User::where('blocked', 1)->pluck('id');
+        $comments = Comment::where('post_id', $post->id)->whereNotIn('user_id', $blocked_users)->orderBy('created_at', 'desc')->paginate(20);
         return view('posts.show', [
         'post' => $post,'comments' => $comments
         ]);
-        }
+    } 
 
 
     public function edit(Post $post) {
-        
-        return view('posts.edit', [
-        'post' => $post
-        ]);
+        if(Auth()->user()->id == $post->user_id ){
+            return view('posts.edit', [
+            'post' => $post
+            ]);
+        } else{
+            abort(403, 'Unauthorized access');
         }
+    }
 
 
     public function create()
@@ -41,7 +46,7 @@ class PostController extends Controller
     public function store(Request $request)
     {
         $request->validate([
-            'text' => 'required']);
+            'text' => 'required|max:1000']);
     
         $post = new Post();
         $post->user_id = Auth()->user()->id;
@@ -49,28 +54,72 @@ class PostController extends Controller
         $post->text = $request->text;
         $post->created_at = now();
         $post->updated_at = now();
+        $user = Auth()->user();
+        if ($post->concern == 1) {
+            $user->concerns = $user->concerns += 1;     
+        }
+        $user->save();
         $post->save();
-    
-        return redirect()->route('posts.index')->with('success', 'Update Successfull');
+        return redirect()->route('posts.index')->with('success', 'Post Created');
     }
     
     public function update(Request $request, Post $post) {
-        $request->validate([
-            'text' => 'required',
-            ]);
-        $post->user_id = Auth()->user()->id;
-        $post->text = $request->text;
-        $post->concern = rand(0,1);
-        $post->updated_at = now();
-        $post->save();
-        return redirect()->route('posts.index')->with('success', 'Update Successfull');
-        }
+        if(Auth()->user()->id == $post->user_id){
+            $request->validate([
+                'text' => 'required|max:1000']);
 
+            $preconcern = $post->concern;
+            $post->user_id = Auth()->user()->id;
+            $post->text = $request->text;
+            $post->concern = rand(0,1);
+            $post->updated_at = now();
+
+            $user = User::find($post->user_id);
+            if($preconcern < $post->concern){
+                $user->concerns = $user->concerns += 1;}
+            elseif($preconcern > $post->concern && $user->concerns > 0){
+                $user->concerns = $user->concerns -= 1;} 
+            else{
+                $user->concerns + 0;
+            }
+            $user->save();
+            $post->save();
+
+            return redirect()->route('posts.index')->with('success', 'Update Successfull');
+        } else{
+            abort(403, 'Unauthorized access');
+        }
+    }
+
+
+        
     
     public function destroy(Post $post) {
-        $post->delete();
-        return redirect()->route('posts.index')->with('success', 'Post deleted successfully');
+        if(Auth()->user()->id == $post->user_id ){
+            if(Auth()->user()->isAdmin == 1){
+                $user = User::find($post->user_id);
+                if($post->concern == 1 && $user->concerns > 0){
+                    $user->concerns = $user->concerns -= 1;
+                }
+                $user->reported -= $post->post_reports;
+                $post->delete();
+                $user->save();
+                return redirect()->back()->with('s uccess', 'Post Deleted Successfully');
+            }
+            else{
+                $user = Auth()->user();
+                if($post->concern == 1 && $user->concerns > 0){
+                    $user->concerns = $user->concerns -= 1;
+                }
+                $user->reported -= $post->post_reports;
+                $post->delete();
+                $user->save();
+                return redirect()->route('posts.index')->with('success', 'Post Deleted Successfully');
+            }
+        } else{
+            abort(403, 'Unauthorized access');
         }
+    }
